requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
google-genai>=0.1.0
psycopg2>=2.9.9
bcrypt==4.1.1
